backend/requirements.txt
fastapi
uvicorn
openai==0.28.1
requests
beautifulsoup4
python-dotenv
